:truck: move PulbicationCard.tsx to a separate folder
